# golang/oapi/openapi.yaml
openapi: 3.0.1
info:
  title: WHZY CRM API
  description: API for managing orders and product
  version: 1.0.0
paths:
  /vendors:
    get:
      tags:
      - whzycrm
      summary: Get all available vendors
      description: Get all available vendor information
      operationId: getVendors
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GetVendorsResponse'
                x-content-type: application/json
        400:
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
...
  /vendor-products/{vendor_id}:
    get:
      tags:
        - whzycrm
      summary: Get list of vendor products
      description: This will return a list of vendor product
      operationId: getVendorProducts
      parameters:
      - name: vendor_id
        in: path
        description: Vendor Id
        required: true
        schema:
          type: integer
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetVendorProductsResponse'
        400:
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        404:
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
  /vendor-product/{vendor_id}/{vendor_product_reference}:
    get:
      tags:
        - whzycrm
      summary: Get all information about a vendor product
      description: This will get product specification, history quotation and its information in orders
      operationId: getVendorProduct
      parameters:
      - name: vendor_id
        in: path
        description: Vendor Id
        required: true
        schema:
          type: integer
      - name: vendor_product_reference
        in: path
        description: Vendor product reference
        required: true
        schema:
          type: string
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetVendorProductResponse'
        400:
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        404:
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
...
components:
  schemas:
    ErrorMessage:
      type: object
      required:
        - message
      properties:
        message:
          type: string
        reference:
          type: string
    Vendor:
      type: object
      required:
        - vendorId
        - name
      properties:
        vendorId:
          type: integer
        name:
          type: string
      example:
        vendorId: 2
        name: ABC factory
...
    Price:
      type: object
      properties:
        currencyId:
          type: integer
        value:
          type: number
        priceTypeId:
          type: integer
      required:
        - currencyId
        - value
        - priceTypeId
    Measure:
      type: object
      properties:
        value:
          type: number
        unit:
          type: string
      required:
        - value
        - unit
    Packing:
      type: object
      properties:
        length:
          $ref: '#/components/schemas/Measure'
        width:
          $ref: '#/components/schemas/Measure'
        height:
          $ref: '#/components/schemas/Measure'
        quantity:
          type: integer
        netWeight:
          $ref: '#/components/schemas/Measure'
        grossWeight:
          $ref: '#/components/schemas/Measure'
    ItemClientQuotationHistory:
      type: object
      properties:
        quotationId:
          type: integer
        date:
          type: string
          format: date
        price:
          $ref: '#/components/schemas/Price'
        moq:
          type: integer
        packing:
          $ref: '#/components/schemas/Packing'
...
    ProductInOrder:
      type: object
      properties:
        clientOrderIRef:
          type: string
        price:
          $ref: '#/components/schemas/Price'
        quantity:
          type: integer
        packing:
          $ref: '#/components/schemas/Packing'
...
    PriceChange:
      type: object
      properties:
        startDate:
          type: string
          format: date
        price:
          $ref: '#/components/schemas/Price'
      required:
        - startDate
        - price
...
    VendorProductSummary:
      type: object
      properties:
        vendorId:
          type: integer
        vendorProductId:
          type: integer
        reference:
          type: string
        description:
          type: string
      required:
      - vendorId
      - vendorProductId
      - reference
      - description
    VendorProduct:
      type: object
      properties:
        vendorId:
          type: integer
        vendorProductId:
          type: integer
        reference:
          type: string
        description:
          type: string
        materialTypeId:
          type: integer
        priceHistory:
          type: array
          items:
            $ref: '#/components/schemas/PriceChange'
        packing:
          $ref: '#/components/schemas/Packing'
      required:
        - vendorId
        - reference
        - priceHistory
        - packing
    GetVendorsResponse:
      type: object
      properties:
        vendors:
          type: array
          items:
            $ref: '#/components/schemas/Vendor'
    GetVendorProductsResponse:
      type: object
      properties:
        vendorProducts:
          type: array
          items:
            $ref: '#/components/schemas/VendorProductSummary'
      required:
      - vendorProducts
    GetVendorProductResponse:
      type: object
      properties:
        vendorProduct:
          $ref: '#/components/schemas/VendorProduct'
        quotations:
          type: array
          items:
            $ref: '#/components/schemas/ItemClientQuotationHistory'
        orders:
          type: array
          items:
            $ref: '#/components/schemas/ProductInOrder'
...